Move REveDataTable to a separate file
